Move string parsing and sequence highlighting into highlight.js
